feat: MyAccont refactoring into smaller components feat: useProfileData hook for fetching new data, updateProfile from zustand/api for updating the data on the page
